Extract mixin.cjs
